+++ apps/web/tailwind.config.ts
@@ -68,6 +68,15 @@ const config: Config = {
       boxShadow: {
         none: 'none',
       },
+      keyframes: {
+        marquee: {
+          '0%': { transform: 'translateX(0%)' },
+          '100%': { transform: 'translateX(-50%)' },
+        },
+      },
+      animation: {
+        marquee: 'marquee 30s linear infinite',
+      },
     },
   },
   plugins: [],